+++ listing/src/main/java/com/listing/listing/repository/ListingRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ListingRepository extends JpaRepository<Listing, Long> {
+public interface ListingRepository extends JpaRepository<Listing, Integer> {
     Page<Listing> findAll(Pageable pageable);
     List<Listing> findAllByUserId(Integer userId, Pageable pageable);
 }
